mv builder to xray
